+++ nix/packages.nix
@@ -31,7 +31,7 @@ let
   addNixVersionFlag = pkg:
     overrideCabal pkg (o: {
       preConfigure = (o.preConfigure or "") + ''
-        if pkg-config --atleast-version 2.4pre nix-store; then
+        if pkg-config --atleast-version 2.4pre nix-store || pkg-config --atleast-version 2.4 nix-store; then
           configureFlags="$configureFlags --flag nix-2_4"
         fi
       '';
@@ -59,6 +59,10 @@ let
 
               # nix-narinfo = self.callPackage ./nix-narinfo.nix { };
 
+              # Must match hercules-ci-cnix-store, which uses `pkgs.nix`.
+              # Nixpkgs may override to a specific series.
+              cachix = super.cachix.override (o: { nix = pkgs.nix; });
+
               hercules-ci-optparse-applicative =
                 super.callPackage ./hercules-ci-optparse-applicative.nix { };
               protolude =
@@ -82,7 +86,7 @@ let
                 let
                   basePkg =
                     callPkg super "hercules-ci-agent" ../hercules-ci-agent {
-                      bdw-gc = pkgs.boehmgc-hercules;
+                      bdw-gc = null; # propagated from Nix instead.
                     };
                   bundledBins = [ pkgs.gnutar pkgs.gzip pkgs.git ] ++ lib.optional pkgs.stdenv.isLinux pkgs.runc;
 
@@ -92,7 +96,7 @@ let
                     (
                       addBuildDepends
                         (enableDWARFDebugging (addNixVersionFlag basePkg))
-                        [ pkgs.makeWrapper pkgs.boost pkgs.boehmgc ]
+                        [ pkgs.makeWrapper pkgs.boost ]
                     )
                     (
                       o:
@@ -163,7 +167,10 @@ let
                 );
               hercules-ci-cnix-expr = addNixVersionFlag
                 (addBuildDepends
-                  (callPkg super "hercules-ci-cnix-expr" ../hercules-ci-cnix-expr { bdw-gc = pkgs.boehmgc-hercules; inherit nix; })
+                  (callPkg super "hercules-ci-cnix-expr" ../hercules-ci-cnix-expr {
+                    bdw-gc = null; # propagated from Nix instead.
+                    inherit nix;
+                  })
                   [
                     # https://github.com/NixOS/nix/pull/4904
                     nlohmann_json