+++ flake.nix
@@ -46,11 +46,11 @@
               nixpkgsSource = nixos-unstable;
               isDevVersion = true;
             };
-            "nixos-unstable-nix_2_4" = {
+            "nixos-unstable-nix_2_3" = {
               nixpkgsSource = nixos-unstable;
               isDevVersion = true;
               overlay = final: prev: {
-                nix = addDebug prev.nix_2_4;
+                nix = addDebug prev.nix_2_3;
               };
             };
             "nixos-unstable-nixUnstable" = {
@@ -65,6 +65,10 @@
             _name: { nixpkgsSource, isDevVersion ? false, overlay ? (_: _: { }) }:
               dimension "System"
                 {
+                  "aarch64-darwin" = {
+                    # eval error (FIXME)
+                    isDevSystem = false;
+                  };
                   "aarch64-linux" = {
                     # shellcheck was broken https://hercules-ci.com/github/hercules-ci/hercules-ci-agent/jobs/826
                     isDevSystem = false;
@@ -170,10 +174,13 @@
                 hercules-ci-cli-nixUnstable =
                   allTargets."nixos-unstable-nixUnstable".${system}.hercules-ci-cli;
 
-                hercules-ci-agent-nix_2_4 =
-                  allTargets."nixos-unstable-nix_2_4".${system}.hercules-ci-agent;
-                hercules-ci-cli-nix_2_4 =
-                  allTargets."nixos-unstable-nix_2_4".${system}.hercules-ci-cli;
+                hercules-ci-agent-nix_2_3 =
+                  allTargets."nixos-unstable-nix_2_3".${system}.hercules-ci-agent;
+                hercules-ci-cli-nix_2_3 =
+                  allTargets."nixos-unstable-nix_2_3".${system}.hercules-ci-cli;
+
+                hercules-ci-agent-nix_2_4 = v.hercules-ci-agent;
+                hercules-ci-cli-nix_2_4 = v.hercules-ci-cli;
               }
           )
           defaultTarget;