+++ nix/make-overlay.nix
@@ -15,27 +15,4 @@ self: pkgs:
 
   inherit (self.hercules-ci-agent-packages) pre-commit-check;
 
-  boehmgc-hercules = (
-    pkgs.boehmgc.override {
-      enableLargeConfig = true;
-    }
-  ).overrideAttrs (
-    attrs: attrs
-      // {
-      patches = (attrs.patches or [ ])
-        ++ [
-        ./boehmgc-8.0.2-min-heap-incr-1mb.patch
-      ]
-      ;
-    }
-  );
-
-  nix = pkgs.nix.override (
-    args:
-    args
-    // {
-      boehmgc = self.boehmgc-hercules;
-    }
-  );
-
 }